deleted dedicated appsync build, merged w main build
