--- front-end-appsync-build/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "front-end",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "aws-amplify": "^1.1.24",
-    "aws-amplify-react": "^2.3.6",
-    "bootstrap": "^4.3.1",
-    "react": "^16.8.5",
-    "react-bootstrap": "^1.0.0-beta.6",
-    "react-datepicker": "^2.3.0",
-    "react-dom": "^16.8.5",
-    "react-scripts": "2.1.8"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "main": "index.js",
-  "devDependencies": {},
-  "author": "",
-  "license": "ISC",
-  "description": ""
-}
