--- front-end-appsync-build/src/App.css
@@ -1,52 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-footer {
-  background-color: #282c34;
-  height:25px;
-  width: 100%;
-  color:  #FFDAB9;
-  text-align: center;
-  bottom: 0;
-  position: fixed;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.text-color-white{
-  color:white;
-}
-
-
-.bg-olive {
-  background-color: #538235;
-}
